% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_forecast.R
\name{auto_forecast}
\alias{auto_forecast}
\title{Automatic forecasting from univariate time series}
\usage{
auto_forecast(data, h = 12, h_cv = 1, window = NULL, acc = "MAE",
  n = 10, level = 0.05, exclude = NULL, test = NULL)
}
\arguments{
\item{data}{univariate time series (ts object).}

\item{h}{forecast horizon.}

\item{h_cv}{forecast horizon to compute accuracy by cross-validation.}

\item{window}{length of the rolling window for cross-validation. If NULL, a rolling window will not be used.}

\item{acc}{accuracy measure to determine the best model. Either "MAE" for mean absolute error, "RMSE" for root mean squared error or "dir" for directional.}

\item{n}{number of bootstrap simulations used to compute both mean forecasts (bagging) and confidence intervals.}

\item{level}{confidence level for prediction intervals.}

\item{exclude}{one or more models to exclude from evaluation, eg. "auto.arima" or c("tbats", "nnetar").}

\item{test}{first observation in the test set. Either a single number or a vector of two integers. If not NULL, overrides cross-validation arguments.}
}
\value{
An object of class list containing the following elements:
\item{fc}{A tibble containing point forecasts and lower/upper limits from the bootstrapped distribution.}
\item{error}{A tibble containing out-of-sample forecast errors.}
\item{acc}{A tibble containing RMSE and MAE values for each model ordered according to the selected criteria.}
\item{dir}{A tibble containing the ordered directional accuracy for each model.}
\item{model}{Character vector with the selected model.}
\item{plot}{A ggplot object with graphical representation of both the point forecasts and confidence intervals.}
}
\description{
This function provides an algorithm to compute the best possible forecast from the available set of univariate time series. Additionaly, bootstrap methods are employed to both refine point forecasts and compute confidence intervals.
}
\details{
Available models are: "auto.arima", "tbats", "ets", "nnetar", "hw", "arfima", "holt", "thetaf", "ses", "meanf", "splinef", "StrucTS", "elm". For more details about these models, see the documentation from forecast and nnfor packages.
}
\examples{
\dontrun{
library(rafa)
auto_forecast(USAccDeaths, test = c(1976,12))
}
}
\author{
J. Renato Leripio
}
