% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafa_boot_pred.R
\name{rafa_boot_pred}
\alias{rafa_boot_pred}
\title{Compute point forecasts and confidence intervals through bootstrapped forecasts.}
\usage{
rafa_boot_pred(data, h, n, level, mod_best_eval)
}
\arguments{
\item{data}{univariate time series (ts object).}

\item{h}{forecast horizon.}

\item{n}{number of bootstrap simulations used to compute both mean forecasts (bagging) and confidence intervals.}

\item{level}{confidence level for prediction intervals.}
}
\value{
A tibble containing point forecasts and lower/upper limits from the bootstrapped distribution.
}
\description{
Compute point forecasts and confidence intervals through bootstrapped forecasts.
}
\examples{
\dontrun{
library(rafa)
}
}
\author{
J. Renato Leripio
}
