% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafa_compute_error.R
\name{rafa_compute_error}
\alias{rafa_compute_error}
\title{Compute forecast errors via cross-validation or train/test sets}
\usage{
rafa_compute_error(data, mod_fun, window = NULL, test = NULL,
  h_cv = 1)
}
\arguments{
\item{data}{univariate time series (ts object).}

\item{mod_fun}{a list of models to be assessed}

\item{window}{length of the rolling window for cross-validation. If NULL, a rolling window will not be used.}

\item{h_cv}{forecast horizon to compute accuracy by cross-validation.}
}
\value{
A list of data frames with errors of the various models used
}
\description{
This function compute forecast errors via cross-validation or train/test sets.
}
\examples{
\dontrun{
library(rafa)
models <- rafa_list_models(exclude = c("nnetar", "hw", "arfima", "holt", "thetaf", "ses", "meanf", "splinef", "StructTS", "elm"))
modelos
errors <- rafa_compute_error(USAccDeaths, mod_fun = models)
errors
}
}
\author{
J. Renato Leripio
}
