% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafa_dir_acc.R
\name{rafa_dir_acc}
\alias{rafa_dir_acc}
\title{Compute the directional accuracy of models}
\usage{
rafa_dir_acc(data, mod_erro)
}
\arguments{
\item{data}{univariate time series (ts object).}

\item{mod_erro}{list of data frames containing the prediction error of various models}
}
\description{
This function compute the directional accuracy of models
}
\examples{
\dontrun{
library(rafa)
models <- rafa_list_models(exclude = c("nnetar", "hw", "arfima", "holt", "thetaf", "ses", "meanf", "splinef", "StructTS", "elm"))
modelos
errors <- rafa_compute_error(USAccDeaths, mod_fun = models)
errors
print(rafa_dir_acc(errors))
}
}
\author{
J. Renato Leripio
}
