% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafa_list_models.R
\name{rafa_list_models}
\alias{rafa_list_models}
\title{Creates list of forecast models to use}
\usage{
rafa_list_models(exclude = NULL, h = 12)
}
\arguments{
\item{exclude}{One or more models to exclude from evaluation, eg. "auto.arima" or c("tbats", "nnetar").}

\item{h}{forecast horizon.}
}
\value{
A list of models.
}
\description{
This function creates a list of models to later use in automatic forecast. It has the same behaviour of the auto_forecast function
}
\details{
Available models are: "auto.arima", "tbats", "ets", "nnetar", "hw", "arfima", "holt", "thetaf", "ses", "meanf", "splinef", "StrucTS". For more details about these models, see the documentation from forecast package.
}
\examples{
\dontrun{
library(rafa)
rafa_list_functions(exclude = NULL)
rafa_list_functions(exclude = "ses")
}
}
\author{
J. Renato Leripio
}
