% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafa_mae_rmse.R
\name{rafa_mae_rmse}
\alias{rafa_mae_rmse}
\title{Compute MAE and RMSE}
\usage{
rafa_mae_rmse(mod_erro)
}
\arguments{
\item{mod_erro}{A list of data frames with errors of the various models used}
}
\value{
An data frame with the MAE and RMSE values of each model used
}
\description{
Compute MAE and RMSE
}
\details{
This function compute MAE and RMSE
}
\examples{
\dontrun{
library(rafa)
models <- rafa_list_models(exclude = c("nnetar", "hw", "arfima", "holt", "thetaf", "ses", "meanf", "splinef", "StructTS", "elm"))
modelos
errors <- rafa_compute_error(USAccDeaths, mod_fun = models)
errors
print(rafa_mae_rmse(errors))
}
}
\author{
J. Renato Leripio
}
