% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafa_order_acc.R
\name{rafa_order_acc}
\alias{rafa_order_acc}
\title{Merge and order the accuracy measures}
\usage{
rafa_order_acc(mod_acc, dir_acc, acc = "MAE")
}
\arguments{
\item{mod_acc}{A data frame with the RMSE and MAE accuracy measures}

\item{dir_acc}{A data frame with directional accuracy measure}

\item{acc}{accuracy measure to determine the best model. Either "MAE" for mean absolute error, "RMSE" for root mean squared error or "dir" for directional.}
}
\value{
An data frame with the MAE and RMSE values of each model used
}
\description{
Merge and order the accuracy measures
}
\details{
This function merge and order the accuracy measures (RMSE, MAE or ACC)
}
\examples{
\dontrun{
library(rafa)
models <- rafa_list_models(exclude = c("nnetar", "hw", "arfima", "holt", "thetaf", "ses", "meanf", "splinef", "StructTS", "elm"))
modelos
errors <- rafa_compute_error(USAccDeaths, mod_fun = models)
errors
print(rafa_mae_rmse(errors))
}
}
\author{
J. Renato Leripio
}
