% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rafa_plot_forecast.R
\name{rafa_plot_forecast}
\alias{rafa_plot_forecast}
\title{Automatic forecasting from univariate time series}
\usage{
rafa_plot_forecast(prev, level, mod_best, n, h)
}
\arguments{
\item{prev}{A tibble containing point forecasts and lower/upper limits from the bootstrapped distribution.}
}
\value{
A ggplot object with graphical representation of both the point forecasts and confidence intervals.
}
\description{
Automatic forecasting from univariate time series
}
\details{
Available models are: "auto.arima", "tbats", "ets", "nnetar", "hw", "arfima", "holt", "thetaf", "ses", "meanf", "splinef", "StrucTS". For more details about these models, see the documentation from forecast package.
}
\examples{
\dontrun{
library(rafa)
}
}
\author{
J. Renato Leripio
}
